made irrlicht.net compile
git-svn-id: svn://svn.code.sf.net/p/irrlicht/code@541 dfc29bdd-3216-0410-991c-e03cc46cb475

+++ source/Irrlicht.NET/IAnimatedMeshSceneNode.cpp
@@ -30,7 +30,7 @@ namespace Scene
 		return getAnimatedMeshSceneNode()->setFrameLoop(begin, end);
 	}
 
-	void IAnimatedMeshSceneNode::set_AnimationSpeed(int framesPerSecond)
+	void IAnimatedMeshSceneNode::set_AnimationSpeed(float framesPerSecond)
 	{
 		getAnimatedMeshSceneNode()->setAnimationSpeed(framesPerSecond);
 	}
